use time.Duration Replace int32

+++ pkg/protocol/codec/global_begin_request_codec.go
@@ -43,8 +43,7 @@ func (c *GlobalBeginRequestCodec) Encode(in interface{}) []byte {
 func (g *GlobalBeginRequestCodec) Decode(in []byte) interface{} {
 	data := message.GlobalBeginRequest{}
 	buf := bytes.NewByteBuffer(in)
-
-	data.Timeout = time.Duration((bytes.ReadUInt32(buf)))
+	data.Timeout = time.Duration(int32(bytes.ReadUInt32(buf)))
 	data.TransactionName = bytes.ReadString16Length(buf)
 
 	return data
